add custom worker API
