+++ .gitignore
@@ -5,3 +5,4 @@ coverage
 test/test-bundle.js
 npm-debug.log
 testdb*
+test/custom-api/worker-bundle.js
