+++ .travis.yml
@@ -30,6 +30,7 @@ env:
   - CLIENT=selenium:firefox:41.0.1 SUITE=5 COMMAND=test
   - CLIENT=selenium:firefox:41.0.1 SUITE=6 COMMAND=test
   - CLIENT=selenium:firefox:41.0.1 SUITE=7 COMMAND=test
+  - COMMAND=test-custom
 
 branches:
   only:
