+++ lib/client/index.js
@@ -1,415 +1,46 @@
 'use strict';
 
-var utils = require('../shared/utils');
-var clientUtils = require('./utils');
-var uuid = require('./../shared/uuid');
-var errors = require('../shared/errors');
-var log = require('debug')('pouchdb:worker:client');
-var preprocessAttachments = clientUtils.preprocessAttachments;
-var encodeArgs = clientUtils.encodeArgs;
-var adapterFun = clientUtils.adapterFun;
+// main script used with a blob-style worker
+
+var extend = require('js-extend').extend;
+var WorkerPouchCore = require('./core');
+var createWorker = require('./create-worker');
 var isSupportedBrowser = require('./is-supported-browser');
 
-var createWorker = require('./create-worker');
-var worker;
-try {
-  worker = createWorker(require('../worker'));
-  worker.addEventListener('error', function (e) {
-    if ('console' in global && 'warn' in console) {
-      console.warn('worker threw an error', e.error);
-    }
-  });
-} catch (e) {
-  if ('console' in global && 'info' in console) {
-    console.info('This browser is not supported by WorkerPouch. ' +
-      'Please use isSupportedBrowser() to check.', e);
-  }
-}
-
-// Implements the PouchDB API for dealing with PouchDB instances over WW
 function WorkerPouch(opts, callback) {
-  var api = this;
 
-  if (typeof opts === 'string') {
-    var slashIdx = utils.lastIndexOf(opts, '/');
-    opts = {
-      url: opts.substring(0, slashIdx),
-      name: opts.substring(slashIdx + 1)
-    };
-  } else {
-    opts = utils.clone(opts);
-  }
-
-  log('constructor called', opts);
-
-  if (!opts.name) {
-    var optsErrMessage = 'Error: you must provide a database name.';
-    console.error(optsErrMessage);
-    return callback(new Error(optsErrMessage));
-  }
-
-  function handleUncaughtError(content) {
+  var worker = window.__pouchdb_global_worker; // cache so there's only one
+  if (!worker) {
     try {
-      api.emit('error', content);
-    } catch (err) {
-      // TODO: it's weird that adapters should have to handle this themselves
-      console.error(
-        'The user\'s map/reduce function threw an uncaught error.\n' +
-        'You can debug this error by doing:\n' +
-        'myDatabase.on(\'error\', function (err) { debugger; });\n' +
-        'Please double-check your map/reduce function.');
-      console.error(content);
-    }
-  }
-
-  function onReceiveMessage(message) {
-    var messageId = message.messageId;
-    var messageType = message.type;
-    var content = message.content;
-
-    if (messageType === 'uncaughtError') {
-      handleUncaughtError(content);
-      return;
-    }
-
-    var cb = api._callbacks[messageId];
-
-    if (!cb) {
-      log('duplicate message (ignoring)', messageId, messageType, content);
-      return;
-    }
-
-    log('receive message', api._instanceId, messageId, messageType, content);
-
-    if (messageType === 'error') {
-      delete api._callbacks[messageId];
-      cb(content);
-    } else if (messageType === 'success') {
-      delete api._callbacks[messageId];
-      cb(null, content);
-    } else { // 'update'
-      api._changesListeners[messageId](content);
-    }
-  }
-
-  function workerListener(e) {
-    if (e.data.id === api._instanceId) {
-      onReceiveMessage(e.data);
-    }
-  }
-
-  function sendMessage(type, args, callback) {
-    var messageId = uuid();
-    log('send message', api._socketId, messageId, type, args);
-    api._callbacks[messageId] = callback;
-    var encodedArgs = encodeArgs(args);
-    worker.postMessage({
-      id: api._instanceId,
-      type: type,
-      messageId: messageId,
-      args: encodedArgs
-    });
-    log('message sent', api._instanceId, messageId);
-  }
-
-  function sendRawMessage(messageId, type, args) {
-    log('send message', api._socketId, messageId, type, args);
-    var encodedArgs = encodeArgs(args);
-    worker.postMessage({
-      id: api._instanceId,
-      type: type,
-      messageId: messageId,
-      args: encodedArgs
-    });
-    log('message sent', api._instanceId, messageId);
-  }
-
-  api.type = function () {
-    return 'socket';
-  };
-
-  api._id = adapterFun('id', function (callback) {
-    sendMessage('id', [], callback);
-  });
-
-  api.compact = adapterFun('compact', function (opts, callback) {
-    if (typeof opts === 'function') {
-      callback = opts;
-      opts = {};
-    }
-    sendMessage('compact', [opts], callback);
-  });
-
-  api._info = function (callback) {
-    sendMessage('info', [], callback);
-  };
-
-  api.get = adapterFun('get', function (id, opts, callback) {
-    if (typeof opts === 'function') {
-      callback = opts;
-      opts = {};
-    }
-    sendMessage('get', [id, opts], callback);
-  });
-
-  // hacky code necessary due to implicit breaking change in
-  // https://github.com/pouchdb/pouchdb/commits/0ddeae6b
-  api._get = function (id, opts, callback) {
-    api.get(id, opts, function (err, doc) {
-      if (err) {
-        return callback(err);
-      }
-      callback(null, {doc: doc});
-    });
-  };
-
-  api.remove =
-    adapterFun('remove', function (docOrId, optsOrRev, opts, callback) {
-      var doc;
-      if (typeof optsOrRev === 'string') {
-        // id, rev, opts, callback style
-        doc = {
-          _id: docOrId,
-          _rev: optsOrRev
-        };
-        if (typeof opts === 'function') {
-          callback = opts;
-          opts = {};
+      worker = createWorker(require('../worker'));
+      worker.addEventListener('error', function (e) {
+        if ('console' in global && 'warn' in console) {
+          console.warn('worker threw an error', e.error);
         }
-      } else {
-        // doc, opts, callback style
-        doc = docOrId;
-        if (typeof optsOrRev === 'function') {
-          callback = optsOrRev;
-          opts = {};
-        } else {
-          callback = opts;
-          opts = optsOrRev;
-        }
-      }
-      var rev = (doc._rev || opts.rev);
-
-      sendMessage('remove', [doc._id, rev], callback);
-  });
-
-  api.getAttachment =
-    adapterFun('getAttachment', function (docId, attachmentId, opts,
-                                                callback) {
-      if (typeof opts === 'function') {
-        callback = opts;
-        opts = {};
-      }
-      sendMessage('getAttachment', [docId, attachmentId, opts], callback);
-  });
-
-  api.removeAttachment =
-    adapterFun('removeAttachment', function (docId, attachmentId, rev,
-                                                   callback) {
-
-      sendMessage('removeAttachment', [docId, attachmentId, rev], callback);
-    });
-
-  // Add the attachment given by blob and its contentType property
-  // to the document with the given id, the revision given by rev, and
-  // add it to the database given by host.
-  api.putAttachment =
-    adapterFun('putAttachment', function (docId, attachmentId, rev, blob,
-                                                type, callback) {
-      if (typeof type === 'function') {
-        callback = type;
-        type = blob;
-        blob = rev;
-        rev = null;
-      }
-      if (typeof type === 'undefined') {
-        type = blob;
-        blob = rev;
-        rev = null;
-      }
-
-      if (typeof blob === 'string') {
-        var binary;
-        try {
-          binary = atob(blob);
-        } catch (err) {
-          // it's not base64-encoded, so throw error
-          return callback(errors.error(errors.BAD_ARG,
-            'Attachments need to be base64 encoded'));
-        }
-        blob = utils.createBlob([utils.binaryStringToArrayBuffer(binary)], {type: type});
-      }
-
-      var args = [docId, attachmentId, rev, blob, type];
-      sendMessage('putAttachment', args, callback);
-    });
-
-  api.put = adapterFun('put', utils.getArguments(function (args) {
-    var temp, temptype, opts;
-    var doc = args.shift();
-    var id = '_id' in doc;
-    var callback = args.pop();
-    if (typeof doc !== 'object' || Array.isArray(doc)) {
-      return callback(errors.error(errors.NOT_AN_OBJECT));
-    }
-
-    doc = utils.clone(doc);
-
-    preprocessAttachments(doc).then(function () {
-      while (true) {
-        temp = args.shift();
-        temptype = typeof temp;
-        if (temptype === "string" && !id) {
-          doc._id = temp;
-          id = true;
-        } else if (temptype === "string" && id && !('_rev' in doc)) {
-          doc._rev = temp;
-        } else if (temptype === "object") {
-          opts = utils.clone(temp);
-        }
-        if (!args.length) {
-          break;
-        }
-      }
-      opts = opts || {};
-
-      sendMessage('put', [doc, opts], callback);
-    }).catch(callback);
-
-  }));
-
-  api.post = adapterFun('post', function (doc, opts, callback) {
-    if (typeof opts === 'function') {
-      callback = opts;
-      opts = {};
-    }
-    opts = utils.clone(opts);
-
-    sendMessage('post', [doc, opts], callback);
-  });
-
-  api._bulkDocs = function (req, opts, callback) {
-    sendMessage('bulkDocs', [req, opts], callback);
-  };
-
-  api._allDocs = function (opts, callback) {
-    if (typeof opts === 'function') {
-      callback = opts;
-      opts = {};
-    }
-    sendMessage('allDocs', [opts], callback);
-  };
-
-  api._changes = function (opts) {
-    opts = utils.clone(opts);
-
-    if (opts.continuous) {
-      var messageId = uuid();
-      api._changesListeners[messageId] = opts.onChange;
-      api._callbacks[messageId] = opts.complete;
-      sendRawMessage(messageId, 'liveChanges', [opts]);
-      return {
-        cancel: function () {
-          sendRawMessage(messageId, 'cancelChanges', []);
-        }
-      };
-    }
-
-    sendMessage('changes', [opts], function (err, res) {
-      if (err) {
-        opts.complete(err);
-        return callback(err);
-      }
-      res.results.forEach(function (change) {
-        opts.onChange(change);
       });
-      if (opts.returnDocs === false || opts.return_docs === false) {
-        res.results = [];
+      window.__pouchdb_global_worker = worker;
+    } catch (e) {
+      if ('console' in global && 'info' in console) {
+        console.info('This browser is not supported by WorkerPouch. ' +
+          'Please use isSupportedBrowser() to check.', e);
       }
-      opts.complete(null, res);
-    });
-  };
-
-  // Given a set of document/revision IDs (given by req), tets the subset of
-  // those that do NOT correspond to revisions stored in the database.
-  // See http://wiki.apache.org/couchdb/HttpPostRevsDiff
-  api.revsDiff = adapterFun('revsDiff', function (req, opts, callback) {
-    // If no options were given, set the callback to be the second parameter
-    if (typeof opts === 'function') {
-      callback = opts;
-      opts = {};
+      return callback(new Error('browser unsupported by worker-pouch'));
     }
-
-    sendMessage('revsDiff', [req, opts], callback);
-  });
-
-  api._query = adapterFun('query', function (fun, opts, callback) {
-    if (typeof opts === 'function') {
-      callback = opts;
-      opts = {};
-    }
-    var funEncoded = fun;
-    if (typeof fun === 'function') {
-      funEncoded = {map: fun};
-    }
-    sendMessage('query', [funEncoded, opts], callback);
-  });
-
-  api._viewCleanup = adapterFun('viewCleanup', function (callback) {
-    sendMessage('viewCleanup', [], callback);
-  });
-
-  api._close = function (callback) {
-    callback();
-  };
-
-  api.destroy = adapterFun('destroy', function (opts, callback) {
-    if (typeof opts === 'function') {
-      callback = opts;
-      opts = {};
-    }
-    sendMessage('destroy', [], function (err, res) {
-      if (err) {
-        api.emit('error', err);
-        return callback(err);
-      }
-      worker.removeEventListener('message', workerListener);
-      api.emit('destroyed');
-      api.constructor.emit('destroyed', api._name);
-      callback(null, res);
-    });
-  });
-
-  api._instanceId = opts.originalName;
-  api._callbacks = {};
-  api._changesListeners = {};
-  api._name = opts.originalName;
-
-  worker.addEventListener('message', workerListener);
-
-  var workerOpts = {
-    name: api._name,
-    auto_compaction: !!opts.auto_compaction
-  };
-  if (opts.revs_limit) {
-    workerOpts.revs_limit = opts.revs_limit;
   }
 
-  sendMessage('createDatabase', [workerOpts], function (err) {
-    if (err) {
-      return callback(err);
-    }
-    callback(null, api);
-  });
+  var _opts = extend({
+    worker: function () { return worker; }
+  }, opts);
 
+  WorkerPouchCore.call(this, _opts, callback);
 }
 
-// WorkerPouch is a valid adapter.
 WorkerPouch.valid = function () {
   return true;
 };
+WorkerPouch.use_prefix = false;
 
 WorkerPouch.isSupportedBrowser = isSupportedBrowser;
-
 
 module.exports = WorkerPouch;
