+++ lib/client/utils.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var utils = require('../shared/utils');
-var log = require('debug')('pouchdb:socket:client');
+var log = require('debug')('pouchdb:worker:client');
 var isBrowser = typeof process === 'undefined' || process.browser;
 
 exports.preprocessAttachments = function preprocessAttachments(doc) {
