+++ lib/workerify/index.js
@@ -28,7 +28,7 @@ function browserifyIt(entry, done) {
 
 function workerify(filename) {
 
-  if (!/\/worker\/index\.js$/.test(filename)) {
+  if (!/\/lib\/worker\/index\.js$/.test(filename)) {
     return through();
   }
   if (processedFiles['$' + filename]) {
