+++ package.json
@@ -29,13 +29,17 @@
     "min": "UGLIFY=1 browserify -s workerPouch -p bundle-collapser/plugin . | ./bin/es3ify.js | derequire | uglifyjs -mc > dist/pouchdb.worker-pouch.min.js",
     "dev": "browserify test/test.js > test/test-bundle.js && npm run dev-server",
     "dev-server": "./bin/dev-server.js",
-    "coverage": "npm test --coverage && istanbul check-coverage --lines 100 --function 100 --statements 100 --branches 100"
+    "coverage": "npm test --coverage && istanbul check-coverage --lines 100 --function 100 --statements 100 --branches 100",
+    "build-custom-worker": "browserify test/custom-api/worker.js --debug > test/custom-api/worker-bundle.js",
+    "test-custom-local": "npm run build-custom-worker && zuul --ui mocha-bdd --local 9000 --no-coverage test/custom-api/test.js",
+    "test-custom": "npm run build-custom-worker && zuul --ui mocha-bdd --phantom --no-coverage test/custom-api/test.js"
   },
   "dependencies": {
     "argsarray": "0.0.1",
     "blob-util": "^1.1.1",
     "debug": "^2.1.3",
     "inherits": "^2.0.1",
+    "js-extend": "^1.0.1",
     "lie": "^2.6.0",
     "pouchdb": "^5.0.0",
     "pouchdb-binary-util": "^1.0.0",
@@ -46,8 +50,8 @@
     "bluebird": "^2.9.24",
     "browserify": "^9.0.8",
     "bundle-collapser": "1.2.1",
-    "chai": "^2.3.0",
-    "chai-as-promised": "^5.1.0",
+    "chai": "^3.5.0",
+    "chai-as-promised": "^5.3.0",
     "corsproxy": "^0.2.14",
     "derequire": "^2.0.0",
     "es3ify": "^0.1.3",
@@ -57,14 +61,15 @@
     "jshint": "2.8.0",
     "mkdirp": "^0.5.0",
     "mocha": "^2.3.3",
-    "phantomjs": "^1.9.7-5",
+    "phantomjs-prebuilt": "^2.1.7",
     "pouchdb-http-proxy": "^0.10.4",
     "request": "^2.36.0",
     "sauce-connect-launcher": "^0.11.1",
     "selenium-standalone": "^4.7.0",
     "uglify-js": "^2.4.13",
     "watchify": "^3.1.0",
-    "wd": "^0.2.21"
+    "wd": "^0.2.21",
+    "zuul": "^3.10.1"
   },
   "browserify": {
     "transform": [
@@ -73,6 +78,8 @@
   },
   "files": [
     "lib",
-    "dist"
+    "dist",
+    "client",
+    "worker"
   ]
 }
