+++ test/index-suite2.html
@@ -25,7 +25,6 @@
 
 <script src='./pouchdb/integration/test.replication.js'></script>
 <script src='./pouchdb/integration/test.sync.js'></script>
-<script src='./pouchdb/integration/browser.info.js'></script>
 <script type="text/javascript" src="./webrunner.js"></script>
 </body>
 </html>
