+++ test/index.html
@@ -36,7 +36,6 @@
 <script src='./pouchdb/integration/test.retry.js'></script>
 <script src='./pouchdb/integration/test.taskqueue.js'></script>
 <script src='./pouchdb/integration/test.design_docs.js'></script>
-<script src='./pouchdb/integration/browser.info.js'></script>
 <script src='./pouchdb/integration/test.issue221.js'></script>
 <script src='./pouchdb/integration/test.issue3179.js'></script>
 <script src='./pouchdb/integration/test.issue3646.js'></script>
