+++ worker/index.js
@@ -1,0 +1,4 @@
+'use strict';
+
+/* jshint worker:true */
+module.exports = require('../lib/worker/core');
